PQ first implementation.

+++ disk-ann/src/main/java/jetbrains/exodus/diskann/VectorReader.java
@@ -3,7 +3,7 @@ package jetbrains.exodus.diskann;
 import it.unimi.dsi.fastutil.longs.LongObjectImmutablePair;
 
 public interface VectorReader {
-    long size();
+    int size();
 
-    LongObjectImmutablePair<float[]> read(long index);
+    LongObjectImmutablePair<float[]> read(int index);
 }
